refactor mines multiplier logic to shared utility
