+++ website/src/lib/server/games/mines.ts
@@ -2,6 +2,7 @@ import { db } from '$lib/server/db';
 import { user } from '$lib/server/db/schema';
 import { eq } from 'drizzle-orm';
 import { redis } from '$lib/server/redis';
+import { calculateMinesMultiplier } from '$lib/utils';
 
 interface MinesSession {
     sessionToken: string;
@@ -111,53 +112,6 @@ export async function minesAutoCashout() {
     }
 }
 
-const getMaxPayout = (bet: number, picks: number, mines: number): number => {
-    const MAX_PAYOUT = 2_000_000;
-    const HIGH_BET_THRESHOLD = 50_000;
-
-    const mineFactor = 1 + (mines / 25);
-    const baseMultiplier = (1.4 + Math.pow(picks, 0.45)) * mineFactor;
-
-    if (bet > HIGH_BET_THRESHOLD) {
-        const betRatio = Math.pow(Math.min(1, (bet - HIGH_BET_THRESHOLD) / (MAX_PAYOUT - HIGH_BET_THRESHOLD)), 1);
-
-        // Direct cap on multiplier for high bets
-        const maxAllowedMultiplier = 1.05 + (picks * 0.1);
-        const highBetMultiplier = Math.min(baseMultiplier, maxAllowedMultiplier) * (1 - (bet / MAX_PAYOUT) * 0.9);
-        const betSizeFactor = Math.max(0.1, 1 - (bet / MAX_PAYOUT) * 0.9);
-        const minMultiplier = (1.1 + (picks * 0.15 * betSizeFactor)) * mineFactor;
-
-        const reducedMultiplier = highBetMultiplier - ((highBetMultiplier - minMultiplier) * betRatio);
-        const payout = Math.min(bet * reducedMultiplier, MAX_PAYOUT);
-
-        return payout;
-    }
-
-    const payout = Math.min(bet * baseMultiplier, MAX_PAYOUT);
-    return payout;
-};
-
 export function calculateMultiplier(picks: number, mines: number, betAmount: number): number {
-    const TOTAL_TILES = 25;
-    const HOUSE_EDGE = 0.05;
-
-    // Calculate probability of winning based on picks and mines
-    let probability = 1;
-    for (let i = 0; i < picks; i++) {
-        probability *= (TOTAL_TILES - mines - i) / (TOTAL_TILES - i);
-    }
-
-    if (probability <= 0) return 1.0;
-
-    // Calculate fair multiplier based on probability and house edge
-    const fairMultiplier = (1 / probability) * (1 - HOUSE_EDGE);
-
-    const rawPayout = fairMultiplier * betAmount;
-    const maxPayout = getMaxPayout(betAmount, picks, mines);
-    const cappedPayout = Math.min(rawPayout, maxPayout);
-    const effectiveMultiplier = cappedPayout / betAmount;
-
-    return Math.max(1.0, Number(effectiveMultiplier.toFixed(2)));
+    return calculateMinesMultiplier(picks, mines, betAmount);
 }
-
-
